Use new style for long options
Long options are now prefixed by `--` and docker complains loudly when
using the previous style which is using only `-`. We consider that most
people will use the latest version of docker and hence only support
completion on the new style.

Also, many options have now both a long and a short option. We may
update completion to support those.