Factorize completion of images/repositories/tags
It makes little sense to complete just tags without considering the
appropriate repository. Therefore, we provide a completion function for
repositories which is used as an helper to complete repositories
with/without tags which in turn is used when we need to complete
images (by hash or by repository/tags).