Enable option stacking for short options without arguments
This enables Zsh to understand commands like "docker run -it
ubuntu". However, by enabling this, this also makes Zsh completes
"docker run -u<tab>" with "docker run -uapprox" which is not valid. The
users have to put the space or the equal sign themselves before trying
to complete. I believe this is a minor annoyance compared to the not let
Zsh be confused when stacking short options.

Signed-off-by: Vincent Bernat <vincent@bernat.im>